refactor: 구체 클래스가 전이되는 정책 다이어그램 표현
