+++ src/main/java/chessgame/domain/state/State.java
@@ -3,6 +3,30 @@ package chessgame.domain.state;
 import chessgame.domain.Board;
 import chessgame.domain.Command;
 
+/**
+ *                                   +-----+-----+
+ *                -------------- <---|   Ready   |-----> ---------------
+ *                |                  +-----+-----+                      |
+ *                |                        |                            |
+ *                |                        |                            |
+ *                |                        |                            |
+ *                |                        |                            |
+ *                |                        |                            |
+ *                |                        v                            |
+ *                |                 +-------+----+                      |
+ *                | ----------- <---|    White   | ---> ----------------v
+ *                |                 +----+-------+                      |
+ *                |                       ^                             |
+ *                |                       |                             |
+ *                |                       v                             |
+ *                |                 +-------+----+                      |
+ *                | ----------- <---|    Black   | ---> ----------------|
+ *                v                 +----+-------+                      v
+ *          +-----+-----+                                        +-----+------+
+ *          |   Error   | <------------------------------------->|    End     |
+ *          +------------+                                      +------------+
+ *
+ */
 public interface State {
     boolean isEnd();
